# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "小说创作应用后端接口文档。",
        "title": "小说创作应用 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/novels": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "分页查询小说列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "novels"
                ],
                "summary": "查询小说列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "当前页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/novel.ListSuccessResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建一条小说记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "novels"
                ],
                "summary": "创建小说",
                "parameters": [
                    {
                        "description": "创建小说请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handler_novel.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/novel.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    }
                }
            }
        },
        "/novels/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据 ID 查询小说详情。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "novels"
                ],
                "summary": "查询小说详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小说 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/novel.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "小说不存在",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据 ID 更新小说记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "novels"
                ],
                "summary": "更新小说",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小说 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新小说请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handler_novel.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/novel.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "小说不存在",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据 ID 真删小说记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "novels"
                ],
                "summary": "删除小说",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小说 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/novel.DeleteSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "小说不存在",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api_handler_novel.CreateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "author_name": {
                    "description": "AuthorName 表示作者名，可以为空。",
                    "type": "string",
                    "example": "爱潜水的乌贼"
                },
                "cover_url": {
                    "description": "CoverURL 表示封面链接，可以为空。",
                    "type": "string",
                    "example": "https://example.com/cover.jpg"
                },
                "description": {
                    "description": "Description 表示简介，可以为空。",
                    "type": "string",
                    "example": "一部关于废土冒险的小说"
                },
                "name": {
                    "description": "Name 表示小说名，不能为空。",
                    "type": "string",
                    "example": "长夜余火"
                },
                "status": {
                    "description": "Status 表示小说状态，只允许连载中或已完结，未传时默认连载中。",
                    "type": "string",
                    "enum": [
                        "连载中",
                        "已完结"
                    ],
                    "example": "连载中"
                },
                "tags": {
                    "description": "Tags 表示标签，可以为空，多个标签使用英文逗号分隔。",
                    "type": "string",
                    "example": "玄幻,冒险"
                }
            }
        },
        "internal_api_handler_novel.UpdateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "author_name": {
                    "description": "AuthorName 表示作者名，可以为空。",
                    "type": "string",
                    "example": "爱潜水的乌贼"
                },
                "cover_url": {
                    "description": "CoverURL 表示封面链接，可以为空。",
                    "type": "string",
                    "example": "https://example.com/cover.jpg"
                },
                "description": {
                    "description": "Description 表示简介，可以为空。",
                    "type": "string",
                    "example": "一部关于废土冒险的小说"
                },
                "name": {
                    "description": "Name 表示小说名，不能为空。",
                    "type": "string",
                    "example": "长夜余火"
                },
                "status": {
                    "description": "Status 表示小说状态，只允许连载中或已完结，未传或空字符串时保留原状态。",
                    "type": "string",
                    "enum": [
                        "连载中",
                        "已完结"
                    ],
                    "example": "连载中"
                },
                "tags": {
                    "description": "Tags 表示标签，可以为空，多个标签使用英文逗号分隔。",
                    "type": "string",
                    "example": "玄幻,冒险"
                }
            }
        },
        "novel.DeleteData": {
            "type": "object",
            "properties": {
                "deleted": {
                    "description": "Deleted 表示是否已经删除。",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "novel.DeleteSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 表示业务响应码，成功固定为 0。",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "Data 表示删除结果。",
                    "allOf": [
                        {
                            "$ref": "#/definitions/novel.DeleteData"
                        }
                    ]
                },
                "message": {
                    "description": "Message 表示响应提示信息。",
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "novel.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 表示错误响应码，使用 HTTP 状态码。",
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "description": "Message 表示用户可理解的错误提示。",
                    "type": "string",
                    "example": "请求参数错误"
                }
            }
        },
        "novel.ListData": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "Items 表示当前页小说列表。",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/novel.NovelData"
                    }
                },
                "page": {
                    "description": "Page 表示当前页码。",
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "description": "PageSize 表示每页数量。",
                    "type": "integer",
                    "example": 20
                },
                "total": {
                    "description": "Total 表示符合条件的小说总数。",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "novel.ListSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 表示业务响应码，成功固定为 0。",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "Data 表示小说分页列表响应数据。",
                    "allOf": [
                        {
                            "$ref": "#/definitions/novel.ListData"
                        }
                    ]
                },
                "message": {
                    "description": "Message 表示响应提示信息。",
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "novel.NovelData": {
            "type": "object",
            "properties": {
                "author_name": {
                    "description": "AuthorName 表示作者名。",
                    "type": "string",
                    "example": "爱潜水的乌贼"
                },
                "cover_url": {
                    "description": "CoverURL 表示封面链接。",
                    "type": "string",
                    "example": "https://example.com/cover.jpg"
                },
                "created_at": {
                    "description": "CreatedAt 表示创建时间。",
                    "type": "string",
                    "example": "2026-06-13T22:00:00+08:00"
                },
                "description": {
                    "description": "Description 表示简介。",
                    "type": "string",
                    "example": "一部关于废土冒险的小说"
                },
                "id": {
                    "description": "ID 表示小说主键 ID。",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "Name 表示小说名。",
                    "type": "string",
                    "example": "长夜余火"
                },
                "status": {
                    "description": "Status 表示小说状态，只允许连载中或已完结。",
                    "type": "string",
                    "enum": [
                        "连载中",
                        "已完结"
                    ],
                    "example": "连载中"
                },
                "tags": {
                    "description": "Tags 表示标签，多个标签使用英文逗号分隔。",
                    "type": "string",
                    "example": "玄幻,冒险"
                },
                "updated_at": {
                    "description": "UpdatedAt 表示更新时间。",
                    "type": "string",
                    "example": "2026-06-13T22:00:00+08:00"
                }
            }
        },
        "novel.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 表示业务响应码，成功固定为 0。",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "Data 表示小说响应数据。",
                    "allOf": [
                        {
                            "$ref": "#/definitions/novel.NovelData"
                        }
                    ]
                },
                "message": {
                    "description": "Message 表示响应提示信息。",
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "输入 Bearer token，例如：Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}